refactor(cloudformation): move get_stack_name to helper module
