+++ fzfaws/cloudformation/create_stack.py
@@ -2,7 +2,6 @@
 import json
 from typing import Any, Dict, Optional, Union
 
-from PyInquirer import prompt
 
 from fzfaws.cloudformation import Cloudformation
 from fzfaws.cloudformation.helper.cloudformationargs import CloudformationArgs
@@ -14,7 +13,8 @@ from fzfaws.cloudformation.helper.file_validation import (
 from fzfaws.cloudformation.helper.paramprocessor import ParamProcessor
 from fzfaws.cloudformation.validate_stack import validate_stack
 from fzfaws.s3 import S3
-from fzfaws.utils import FileLoader, Pyfzf, StackNameValidator, prompt_style
+from fzfaws.utils import FileLoader, Pyfzf
+from fzfaws.cloudformation.helper import get_stack_name
 
 
 def create_stack(
@@ -207,19 +207,3 @@ def construct_s3_creation_args(
     }
 
     return cloudformation_args
-
-
-def get_stack_name() -> str:
-    """Get user to input the stack name."""
-    questions = [
-        {
-            "type": "input",
-            "name": "answer",
-            "message": "StackName",
-            "validate": StackNameValidator,
-        }
-    ]
-    result = prompt(questions, style=prompt_style)
-    if not result:
-        raise KeyboardInterrupt
-    return result.get("answer", "")
